feat: add AdoptOpenJDK tap
Since it is become more and more necessary to install Java versions
other than 8, this tap is much better suited in lieu of the old
homebrew/cask-versions which only contained adoptopenjdk8.

Thus also installs adoptopenjdk11 -- time to keep up with the times.
Java 1.8 remains the default, though, for the foreseeable future.