feat: register Homebrew build JDKs properly
Bottles like openjdk and openjdk@ are built by Homebrew, as opposed to
oracle-jdk, microsoft-openjdk, adoptopenjdk, temurin, zulu...

They do not link to /Library/Java/JavaVirtualMachines when installing
(maybe that is why sudo is not required, which is a positive).

The downside is that to register the JDKs with jenv the Homebrew path
must be used, which is architecture dependent. A small price to pay?

Note that as of this commit jenv is bugged and does not create all the
proper aliases for a JDK version. See the link below for more details:

- https://github.com/jenv/jenv/issues/373